Restore 'dx-chart-core' tests
