+++ packages/dx-chart-core/src/plugins/scale/computeds.test.ts
@@ -1,24 +1,10 @@
 import {
-  defaultDomains, addDomain, extendDomains, buildScales, scaleLinear, scaleBand,
+  defaultDomains, addDomain, extendDomains, buildScales,
 } from './computeds';
+import {
+  scaleLinear, scaleBand,
+} from '../../utils/scale';
 import { ARGUMENT_DOMAIN, VALUE_DOMAIN } from '../../constants';
-
-jest.mock('d3-scale', () => ({
-  scaleLinear: () => ({ tag: 'scale-linear' }),
-  scaleBand: () => {
-    const ret = { tag: 'scale-band' } as any;
-    ret.paddingInner = (value) => {
-      ret.inner = value;
-      return ret;
-    };
-    ret.paddingOuter = (value) => {
-      ret.outer = value;
-      return ret;
-    };
-    ret.bandwidth = () => 0;
-    return ret;
-  },
-}));
 
 describe('Scale', () => {
   describe('defaultDomains', () => {
@@ -63,23 +49,6 @@ describe('Scale', () => {
         'domain-1': { tag: '1' },
         'domain-2': { tag: '2' },
         'test-domain': { domain: [], factory, isDiscrete: true },
-      });
-    });
-  });
-
-  describe('default scales', () => {
-    it('should provide linear scale', () => {
-      expect(scaleLinear()).toEqual({ tag: 'scale-linear' });
-    });
-
-    it('should provide band scale', () => {
-      expect(scaleBand()).toEqual({
-        tag: 'scale-band',
-        inner: 0.3,
-        outer: 0.15,
-        paddingInner: expect.any(Function),
-        paddingOuter: expect.any(Function),
-        bandwidth: expect.any(Function),
       });
     });
   });
@@ -323,7 +292,10 @@ describe('Scale', () => {
         [ARGUMENT_DOMAIN]: { domain: 'test-domain-1', factory: () => mockScale1 },
         [VALUE_DOMAIN]: { domain: 'test-domain-2', factory: () => mockScale2 },
         'test-domain': { domain: 'test-domain-3', factory: () => mockScale3 },
-      } as any, { width: 400, height: 300 });
+      }, {
+        [ARGUMENT_DOMAIN]: [0, 400],
+        [VALUE_DOMAIN]: [300, 0]
+      });
 
       expect(scales).toEqual({
         [ARGUMENT_DOMAIN]: mockScale1,
